// src/shared/types/astro-planets.ts
export enum ASTRO_PLANETS {
  SUN = 'Sun',
  MOON = 'Moon',
  MERCURY = 'Mercury',
  VENUS = 'Venus',
  MARS = 'Mars',
  JUPITER = 'Jupiter',
  SATURN = 'Saturn',
  URANUS = 'Uranus',
  NEPTUNE = 'Neptune',
  PLUTO = 'Pluto',

  PROSERPINA = 'Proserpina',
  CHIRON = 'Chiron',
  LILITH = 'Lilith',
  RAHU = 'Rahu',
  KETU = 'Ketu', // Южный узел будет вычисляться как Rahu + 180°
  FORTUNA = 'Fortuna',
  SELENA = 'Selena',
}

export interface IPlanet {
  name: ASTRO_PLANETS
  longitude: number
  latitude: number
  distance: number
  speed: number
  isRetrograde: boolean
}